acmp1325 solution

+++ c/acmp/acmp1325.cpp
@@ -9,7 +9,7 @@ int main(int argc, char **argv) {
 	int n, q;
 	cin >> n >> q;
 	int sz = max(q, n*6)+1;
-	vector<vector<double>> vv(2, vector<double>(sz));
+	vector<vector<long double>> vv(2, vector<long double>(sz));
 	vv[0][0] = 1;
 	for (int i=0; i<n; i++) {
 		int l = i&1;
@@ -19,7 +19,7 @@ int main(int argc, char **argv) {
 			for (int d=1; d<7; d++)
 				vv[nl][j+d] += vv[l][j];
 	}
-	double sm = accumulate(vv[n&1].begin(), vv[n&1].end(), 0.0);
+	long double sm = accumulate(vv[n&1].begin(), vv[n&1].end(), (long double)0.0);
 	cout << vv[n&1][q]/sm << endl;
 	return 0;
 }
